src/backend/pyproject.toml
[tool.poetry]
name = "django_slack_remover"
version = "0.1.0"
description = ""
authors = ["Adam <adamsmelkowski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.2"
slackclient = "^2.9.1"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
ipython = "^7.18.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.5.4"
black = "^20.8b1"
flake8 = "^3.8.4"
ipdb = "^0.13.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
